up JUnit test

+++ pig_farm/src/main/java/vn/codegym/pig_farm/controller/EmployeeRestController.java
@@ -9,16 +9,18 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import vn.codegym.pig_farm.entity.Employee;
+import vn.codegym.pig_farm.projection.IEmployeeProjection;
 import vn.codegym.pig_farm.service.IEmployeeService;
 
+import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @RestController
 @CrossOrigin
 @RequestMapping("/employee")
 public class EmployeeRestController {
-
-
+    
     @Autowired
     private IEmployeeService iEmployeeService;
 
@@ -31,12 +33,12 @@ public class EmployeeRestController {
      * @return if success status 2xx else if error status 4xx
      */
     @GetMapping("/searchList")
-    public ResponseEntity<Page<Employee>> getAllListEmployeePaginationAndSearch(@PageableDefault(value = 2) Pageable pageable,
-                                              @RequestParam Optional<String> name,
-                                              @RequestParam Optional<String> idCard) {
+    public ResponseEntity<Page<IEmployeeProjection>> getAllListEmployeePaginationAndSearch(@PageableDefault(value = 2) Pageable pageable,
+                                                                                           @RequestParam Optional<String> name,
+                                                                                           @RequestParam Optional<String> idCard) {
         String keywordIdCard = idCard.orElse("");
         String keywordName = name.orElse("");
-        Page<Employee> employeePage = iEmployeeService.getAllEmployeePaginationAndSearch(keywordName,keywordIdCard, pageable);
+        Page<IEmployeeProjection> employeePage = iEmployeeService.getAllEmployeePaginationAndSearch(keywordName,keywordIdCard, pageable);
         if (employeePage.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
@@ -51,7 +53,15 @@ public class EmployeeRestController {
      */
     @PatchMapping("/delete/{id}")
     public ResponseEntity<Employee> deleteEmployeeById(@PathVariable Integer id){
-        iEmployeeService.deleteEmployee(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        List<Employee> employeeList = iEmployeeService.getAllEmployee();
+        for (Employee employee: employeeList) {
+            if(Objects.equals(id, employee.getId())){
+                iEmployeeService.deleteEmployee(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
+
+
 }
